Refactor DynamicResourceHandlerExtensions
