+++ src/CommunityToolkit.Maui.Markup/ElementExtensions.cs
@@ -56,38 +56,6 @@ public static class ElementExtensions
 	}
 
 	/// <summary>
-	/// Set Dynamic Resource
-	/// </summary>
-	/// <typeparam name="TDynamicResourceHandler"></typeparam>
-	/// <param name="dynamicResourceHandler"></param>
-	/// <param name="property"></param>
-	/// <param name="key"></param>
-	/// <returns>Layout with added Dynamic Resource</returns>
-	public static TDynamicResourceHandler DynamicResource<TDynamicResourceHandler>(this TDynamicResourceHandler dynamicResourceHandler, BindableProperty property, string key) where TDynamicResourceHandler : IDynamicResourceHandler
-	{
-		dynamicResourceHandler.SetDynamicResource(property, key);
-
-		return dynamicResourceHandler;
-	}
-
-	/// <summary>
-	/// Set Dynamic Resource
-	/// </summary>
-	/// <typeparam name="TDynamicResourceHandler"></typeparam>
-	/// <param name="dynamicResourceHandler"></param>
-	/// <param name="resources"></param>
-	/// <returns>Layout with added Dynamic Resource</returns>
-	public static TDynamicResourceHandler DynamicResources<TDynamicResourceHandler>(this TDynamicResourceHandler dynamicResourceHandler, params (BindableProperty property, string key)[] resources) where TDynamicResourceHandler : IDynamicResourceHandler
-	{
-		foreach (var (property, key) in resources)
-		{
-			dynamicResourceHandler.DynamicResource(property, key);
-		}
-
-		return dynamicResourceHandler;
-	}
-
-	/// <summary>
 	/// Remove Dynamic Resource
 	/// </summary>
 	/// <typeparam name="TBindable"></typeparam>
